# wgengine: instrument with usermetrics