This PR adds metrics in quite performance sensitive part of the code, please help me figure if this is approach is reasonable.

Updates tailscale/corp#22075