@kradalby 2024-10-07
I've ran som of the benchmarks in `magicsock` to compare:

`main`:
```
16:03:17 ❯ go test --bench=. --benchtime=20s
goos: darwin
goarch: arm64
pkg: tailscale.com/wgengine/magicsock
cpu: Apple M2
BenchmarkReceiveFrom-8                   2872041              8667 ns/op              52 B/op          2 allocs/op
--- BENCH: BenchmarkReceiveFrom-8
    peermtu.go:95: [v1] magicsock: peermtu: peer MTU status is false
    magicsock.go:519: magicsock: disco key = d:cc03cc74f66a839e
    magicsock_test.go:1291: endpoints: [{"127.0.0.1:61851" "local"}]
    peermtu.go:95: [v1] magicsock: peermtu: peer MTU status is false
    magicsock.go:519: magicsock: disco key = d:d9507963e2286394
    magicsock_test.go:1291: endpoints: [{"127.0.0.1:57639" "local"}]
    peermtu.go:95: [v1] magicsock: peermtu: peer MTU status is false
    magicsock.go:519: magicsock: disco key = d:4282f90b28964f64
    magicsock_test.go:1291: endpoints: [{"127.0.0.1:59832" "local"}]
    peermtu.go:95: [v1] magicsock: peermtu: peer MTU status is false
        ... [output truncated]
BenchmarkReceiveFrom_Native-8            2749870              8722 ns/op               4 B/op          1 allocs/op
PASS
ok      tailscale.com/wgengine/magicsock        70.035s
```

this pr:
```
16:01:45 ❯ go test --bench=. --benchtime=20s
goos: darwin
goarch: arm64
pkg: tailscale.com/wgengine/magicsock
cpu: Apple M2
BenchmarkReceiveFrom-8                   2742268              8676 ns/op              52 B/op          2 allocs/op
--- BENCH: BenchmarkReceiveFrom-8
    peermtu.go:95: [v1] magicsock: peermtu: peer MTU status is false
    magicsock.go:573: magicsock: disco key = d:18326a8e8b282f9c
    magicsock_test.go:1377: endpoints: [{"127.0.0.1:65175" "local"}]
    peermtu.go:95: [v1] magicsock: peermtu: peer MTU status is false
    magicsock.go:573: magicsock: disco key = d:7528b7275d891e38
    magicsock_test.go:1377: endpoints: [{"127.0.0.1:62192" "local"}]
    peermtu.go:95: [v1] magicsock: peermtu: peer MTU status is false
    magicsock.go:573: magicsock: disco key = d:17093ca50615b37f
    magicsock_test.go:1377: endpoints: [{"127.0.0.1:62019" "local"}]
    peermtu.go:95: [v1] magicsock: peermtu: peer MTU status is false
        ... [output truncated]
BenchmarkReceiveFrom_Native-8            2752149              8694 ns/op               4 B/op          1 allocs/op
PASS
ok      tailscale.com/wgengine/magicsock        69.130s
```

This benchmark tests receiving, it does wire up the metrics and they don't seem to impact on this kind of synthetic benchmark.